Big updates to bibframe2ef and postConversionTransform
bibframe2ef - place is now actually selected in the SPARQL. Added inLanguage to solr JSON. Added notes on what common fields we want in solr's schema.xml. Built standalone node script that turns JSON-LD results into solr JSONs.

postConversionTransform - When building a blank node from a URL, check if the hash character is present, and generate id based on that data, so we can avoid collisions when it's not there.